-COnfigured the project to automatically set up a database when a database is not present -Started documentind endpoints -introduced the init_db method
